moved audio slicing to prepare files.py and renamed sliceaudios to extract features
